Using scoped services
Add GameDataLogger for logging game info in CreateGameEndpoint

A new class `GameDataLogger` was added in the `GameStore.Api.Data` namespace. This class logs information about each game retrieved from `GameStoreData`.

The `CreateGameEndpoint` class was modified to include `GameDataLogger` as a parameter in the `MapCreateGame` method, allowing the endpoint to log game information after a new game is created.

The `PrintGames` method of `GameDataLogger` is called after a new game is added in the `CreateGameEndpoint` class.

In `Program.cs`, the `GameDataLogger` service was added to the dependency injection container using `AddTransient`. Additionally, the `GameStoreData` service registration was changed from `AddTransient` to `AddScoped`.